euth: move offlinephases validators and widget to contrib
